Change(checkpoint): Reintroduced model crate
