+++ crates/commands/src/file_operations.rs
