+++ crates/commands/src/folder_management.rs
