+++ crates/commands/src/library_management.rs
