+++ crates/commands/src/thumbnail_service.rs
