+++ crates/commands/src/util.rs
