+++ crates/config/Cargo.toml
@@ -4,6 +4,10 @@ edition.workspace = true
 version.workspace = true
 
 [dependencies]
+tracing = { workspace = true }
+serde = { workspace = true }
+anyhow = { workspace = true }
+model = { workspace = true }
 
 [lints]
 workspace = true
