+++ crates/model/src/commands/watched_folders.rs
