+++ crates/model/src/services/mod.rs
@@ -1,3 +1,4 @@
+pub mod decorations;
 pub mod file;
 pub mod folder;
 pub mod tag;
