+++ crates/services/Cargo.toml
@@ -13,5 +13,6 @@ serde = { workspace = true }
 entity = { workspace = true }
 chrono = { workspace = true }
 repositories = { workspace = true }
+
 [lints]
 workspace = true
